Refactor to css module
