+++ app/components/buttons/ToggleSwitch/ToggleSwitch.module.css
@@ -1,0 +1,47 @@
+.toggleSwitch {
+  display: inline-block;
+  vertical-align: middle;
+  margin-right: 0.5em;
+
+  margin-bottom: 10px;
+  width: 85px;
+}
+
+.toggleSwitchEnabled,
+.toggleSwitchDisabled {
+  display: block;
+  width: 50px;
+  height: 20px;
+  border-radius: 3px;
+  transition: all 200ms ease-in-out;
+  background-color: var(--toggle-switch-bg);
+  float: left;
+}
+
+.toggleSwitchEnabled:hover {
+  width: 47px;
+  padding-left: 3px;
+}
+
+.toggleSwitchDisabled:hover {
+  width: 47px;
+  padding-right: 3px;
+}
+
+.toggleSwitchKnobEnabled,
+.toggleSwitchKnobDisabled {
+  width: 25px;
+  height: 100%;
+  border-radius: 3px;
+  box-shadow: 0 3px 8px rgba(12, 30, 62, 0.21);
+  transition: all 200ms ease-in-out;
+}
+  
+.toggleSwitchKnobEnabled {
+  background-color: var(--input-color);
+}
+
+.toggleSwitchKnobDisabled {
+  float: right;
+  background-color: var(--toggle-switch-knob-bg);
+}
